+++ compiler/src/org.graalvm.compiler.truffle.test/src/org/graalvm/compiler/truffle/test/nodes/StoreLocalTestNode.java
@@ -42,7 +42,7 @@ public class StoreLocalTestNode extends AbstractTestNode {
 
     @Override
     public int execute(VirtualFrame frame) {
-        slot.setKind(FrameSlotKind.Int);
+        frame.getFrameDescriptor().setFrameSlotKind(slot, FrameSlotKind.Int);
         int value = valueNode.execute(frame);
         frame.setInt(slot, value);
         return value;