+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/ReturnTypeSpecializationTest.java
@@ -61,10 +61,10 @@ public class ReturnTypeSpecializationTest {
         FrameSlot slot = frameDescriptor.addFrameSlot("localVar", FrameSlotKind.Int);
         TestRootNode rootNode = new TestRootNode(frameDescriptor, new IntAssignLocal(slot, new StringTestChildNode()), new IntReadLocal(slot));
         CallTarget target = runtime.createCallTarget(rootNode);
-        Assert.assertEquals(FrameSlotKind.Int, slot.getKind());
+        Assert.assertEquals(FrameSlotKind.Int, frameDescriptor.getFrameSlotKind(slot));
         Object result = target.call();
         Assert.assertEquals("42", result);
-        Assert.assertEquals(FrameSlotKind.Object, slot.getKind());
+        Assert.assertEquals(FrameSlotKind.Object, frameDescriptor.getFrameSlotKind(slot));
     }
 
     class TestRootNode extends RootNode {
@@ -134,7 +134,7 @@ public class ReturnTypeSpecializationTest {
                 int result = value.executeInt(frame);
                 frame.setInt(slot, result);
             } catch (UnexpectedResultException e) {
-                slot.setKind(FrameSlotKind.Object);
+                frame.getFrameDescriptor().setFrameSlotKind(slot, FrameSlotKind.Object);
                 frame.setObject(slot, e.getResult());
                 replace(new ObjectAssignLocal(slot, value));
             }
@@ -154,7 +154,7 @@ public class ReturnTypeSpecializationTest {
         @Override
         Object execute(VirtualFrame frame) {
             Object o = value.execute(frame);
-            slot.setKind(FrameSlotKind.Object);
+            frame.getFrameDescriptor().setFrameSlotKind(slot, FrameSlotKind.Object);
             frame.setObject(slot, o);
             return null;
         }